# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "whell-laravel-api-documentation",
        "version": "v1.0.0"
    },
    "paths": {
...
        "/api/admin/system/job/{id}": {
            "get": {
                "tags": [
                    "定时任务"
                ],
                "summary": "查询定时任务",
                "description": "查询定时任务",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\System\\JobController::get",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "定时任务id",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobData"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "定时任务"
                ],
                "summary": "修改定时任务状态",
                "description": "修改定时任务状态",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\System\\JobController::modify",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "定时任务id",
                        "required": true
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "当前状态",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "description": "成功状态 1刷新成功 0刷新失败",
                                    "type": "integer",
                                    "example": 1
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/system/job": {
            "get": {
                "tags": [
                    "定时任务"
                ],
                "summary": "查询定时任务列表",
                "description": "查询定时任务列表",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\System\\JobController::search",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "conditions",
                        "in": "query",
                        "description": "条件",
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "名称",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "从第几个开始取",
                        "schema": {
                            "type": "int",
                            "default": "0"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "取几个",
                        "schema": {
                            "type": "int",
                            "default": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/JobData"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "paging": {
                                            "$ref": "#/components/schemas/paging"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/system/job_log": {
            "get": {
                "tags": [
                    "定时任务日志"
                ],
                "summary": "查询定时任务日志列表",
                "description": "查询定时任务日志列表",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\System\\JobLogController::search",
                "parameters": [
                    {
                        "name": "accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/accept"
                        }
                    },
                    {
                        "name": "jobId",
                        "in": "query",
                        "description": "任务id",
                        "required": true
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "从第几个开始取",
                        "schema": {
                            "type": "int",
                            "default": "0"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "取几个",
                        "schema": {
                            "type": "int",
                            "default": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#/components/schemas/JobLogData"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "paging": {
                                            "$ref": "#/components/schemas/paging"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/system/verification_code": {
            "get": {
                "tags": [
                    "验证码"
                ],
                "summary": "登陆验证码",
                "description": "登陆验证码",
                "operationId": "App\\Http\\Controllers\\Api\\Admin\\System\\VerificationCodeController::search",
                "parameters": [
                    {
                        "name": "config",
                        "in": "query",
                        "description": "验证码配置",
                        "schema": {
                            "type": "string",
                            "default": "flat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sensitive": {
                                            "description": "是否开启严格模式",
                                            "type": "bool",
                                            "example": false
                                        },
                                        "key": {
                                            "description": "验证码key",
                                            "type": "string",
                                            "example": "eyJpdiI6ImZvbUlmRUJKajFhVDdCS3gyNHo..."
                                        },
                                        "img": {
                                            "description": "验证码base64",
                                            "type": "string",
                                            "example": "data:image/png;base64,iVBORw0KGg..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "JobData": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "名称",
                        "type": "string",
                        "example": "testJob1"
                    },
                    "expression": {
                        "description": "任务执行表达式",
                        "type": "string",
                        "example": "*\\1 * * * *"
                    },
                    "class": {
                        "description": "执行任务类名",
                        "type": "string",
                        "example": "App\\Models\\Job\\Job\\TestJob"
                    },
                    "args": {
                        "description": "参数",
                        "type": "string",
                        "example": "[]"
                    },
                    "nextExecutionTime": {
                        "description": "下次执行时间",
                        "type": "string",
                        "example": "2021-03-01 16:23:00"
                    },
                    "lastExecutionTime": {
                        "description": "上次执行时间",
                        "type": "string",
                        "example": null
                    },
                    "status": {
                        "description": "状态 1启用 0禁用",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "JobLogData": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "example": 1
                    },
                    "parentId": {
                        "description": "任务id",
                        "type": "integer",
                        "example": 1
                    },
                    "args": {
                        "description": "参数",
                        "type": "string",
                        "example": "[]"
                    },
                    "resulted": {
                        "description": "任务执行结果 1成功 0失败",
                        "type": "integer",
                        "example": 1
                    },
                    "trace": {
                        "description": "异常信息",
                        "type": "string",
                        "example": "执行失败:test error:SQLSTATE[HY000]: General error: 1364 Field 'trace' doesn't have a default value (SQL: insert into `job_log` (`parentId`, `name`, `expression`, `class`, `args`, `resulted`, `costTime`, `createTime`) values (1, test, *\\1 * * * *, App\\Models\\Job\\Job\\TestJob, [], 1, 0.016108989715576, 2020-12-15 20:46:07))line:671file:/Users/ogg/www/whell-laravel/vendor/laravel/framework/src/Illuminate/Database/Connection.php"
                    },
                    "costTime": {
                        "description": "执行花费时间",
                        "type": "string",
                        "example": "0.029523849487305"
                    },
                    "createTime": {
                        "description": "任务开始执行时间",
                        "type": "string",
                        "example": "2021-01-03T10:43:01.000000Z"
                    },
                    "name": {
                        "description": "任务名称",
                        "type": "string",
                        "example": "testJob1"
                    },
                    "expression": {
                        "description": "任务执行表达式",
                        "type": "string",
                        "example": "*\\1 * * * *"
                    },
                    "class": {
                        "description": "执行任务类名",
                        "type": "string",
                        "example": "App\\Models\\Job\\Job\\TestJob"
                    }
                },
                "type": "object"
            },
...
            "accept": {
                "default": "application/whell.api+json"
            },
            "paging": {
                "properties": {
                    "total": {
                        "description": "总数",
                        "type": "integer",
                        "example": 1
                    },
                    "offset": {
                        "description": "从第几个开始取",
                        "type": "integer",
                        "example": 0
                    },
                    "limit": {
                        "description": "取了几个",
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}